Separate handlers for pages and REST APIs
- These are related to generating data.
- Fill in empty tags in GenDataParam
- Display Data Mold Web UI URL when running
